requirements-celery.txt
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
flower>=2.0.1 